WRP-2751: fix order of acceptability membership field update
First execute the refset ID removals then the additions. This will
prevent unwanted elimination of acceptability membership of description
in case of a request, which deletes/inactivates a members and reactives
and changes the acceptability of another member in the same time.

Added test case to reproduce and verify issue.