WRP-2927: Extend the set of concepts to be displayed in a merge review
The method of computation is changed to use storage keys; this is
required because some of the changes can not be differentiated just by
looking at the concept identifiers (eg. when a concept is removed on a
branch and a completely different concept takes its place due to
identifier re-use). Added concepts are also considered, as some existing
concepts can appear as such in the review results, when rebasing a stale
task branch.