SO-1782: support partial document loading by set of field names
This is a work in progress feature.
Partial revision documents will throw IllegalStateException when trying
to use it in Collection which requires hashCode/equals to work properly.
Partial document loading try to use Lucene DocValues and
CollectorManager to speed up the query+fetch phase (especially the
fetch) when we require all matching documents for a query to be loaded
(for example we need all concept IDs for a matching ECL expression).