APDS-124 support for re-index mode when term. server is starting
- removed legacy codesystems re-index code from CDOBasedRepository
- allowing maintenance startup: when started with -Dsnowowl.reindex=true
allowing the boot process to finish so that the healing process of
re-indexing can be kicked off
- when starting in re-index mode, defaults in SnowOwlConfiguration are
modified to values desired for the re-index process
- when not started in re-index mode, the term server boot process fails
fast in the presence of inconsistent repositories, preventing the rest
app. to connect, but still allowing Virgo to finish booting and allowing
access to OSGi console.